generalized rebase process
