+++ go/libraries/doltcore/schema/tag.go
@@ -53,6 +53,7 @@ const (
 
 var randGen = rand.New(rand.NewSource(time.Now().UnixNano()))
 
+// TODO: what's the best, most predictable way to deterministically gen tags?
 func AutoGenerateTag(ss *SuperSchema) uint64 {
 	var maxTagVal uint64 = 128 * 128
 
@@ -78,6 +79,7 @@ func AutoGenerateTag(ss *SuperSchema) uint64 {
 	return randTag
 }
 
+// TODO: how do we want to use this function, what behavior do we want?
 func ResolveTagConflicts(ss *SuperSchema, sch Schema) (Schema, error) {
 	cc, _ := NewColCollection()
 	err := sch.GetAllCols().Iter(func(tag uint64, col Column) (stop bool, err error) {
